test: add Miri tests, fix a stacked borrows violation (#32)
This branch adds support for running Miri tests against `thingbuf`. The
primary reason to use Miri is for testing the wait queue linked list
implementation. Other unsafe code in `thingbuf` is primarily unsafe due
to potential concurrent access issues (which are tested for using
`loom`), rather than doing anything particularly weird with raw pointers
and lifetimes (for which we would want to use Miri).

In order to actually test the implementation using Miri, I wrote a few
tests that exercise basic functionality for the wait queue. 

In the process, I fixed two issues:
- A potential stacked borrows violation in the `enqueue` operation due
to creating a raw pointer to the node being enqueued while there is also
an `&mut` reference to it. This code doesn't currently result in
anything bad happening, as the `&mut` ref doesn't outlive the stack
frame in which the raw ptr is created, but it's a stacked borrows
violation and is technically unsound. This was solved by re-ordering
some of the code so that the raw pointer to the node being enqueued is
not created until we are finished mutating it through the mut ref. 
- Some debug assertions didn't take into account the potential for
`compare_exchange_weak` to fail spuriously with the expected value, and
asserted it was not that value. Apparently miri causes spurious
`compare_exchange_weak` failures (which was surprising to me) and
triggered those assertions. I'm surprised they weren't also triggered
under loom!

There's now a bash script for running Miri locally, and I've added a CI
job to run Miri tests.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>